# boleto-api-python/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Boleto-API (Python)",
    "description": "Gateway de cobrança multi-banco (C6/Sicoob) + proxy ao engine brcobrança (Ruby).\n\n**Produto standalone** consumido por múltiplos sistemas; escopo por `tenant_id`.\n\n**Push de eventos (saída, não é um path desta API):** ao receber o webhook do banco, o gateway envia o evento normalizado (`WebhookEvent`) por `POST` ao consumidor dono do tenant, assinado em `X-Signature: sha256=<hmac_sha256(secret, raw_body)>`. Destino por tenant via `SUB__<tenant>__URL/SECRET`, com fallback global `EVENT_WEBHOOK_URL/SECRET`.\n\n**Respostas `dict` (passthrough):** rotas Pix/Pix Automático/extrato/conciliação devolvem o corpo do banco como veio (padrão BACEN/documentação do banco), sem re-tipagem — o contrato é o do banco; o gateway normaliza autenticação, multi-tenant e erros.",
    "version": "0.6.0"
  },
  "paths": {
...
    "/pix": {
      "post": {
        "tags": [
          "pix"
        ],
        "summary": "Criar",
        "description": "Cria cobrança Pix: cob imediata (default) ou cobv se `data_vencimento`.",
        "operationId": "criar_pix_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer bapi_... (token do /credenciais)",
              "title": "Authorization"
            },
            "description": "Bearer bapi_... (token do /credenciais)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PixCobrancaIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PixCobrancaOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "pix"
        ],
        "summary": "Listar",
        "description": "Lista cobranças do período (RFC3339). `vencimento=true` lista cobv.",
        "operationId": "listar_pix_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "inicio",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Inicio"
            }
          },
          {
            "name": "fim",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Fim"
            }
          },
          {
            "name": "vencimento",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Vencimento"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Provider",
              "default": "c6"
            }
          },
          {
            "name": "X-Bank-Credentials",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Credenciais do banco (JSON em base64) — só memória, nunca persistidas. Fallback: cofre VAULT__*.",
              "title": "X-Bank-Credentials"
            },
            "description": "Credenciais do banco (JSON em base64) — só memória, nunca persistidas. Fallback: cofre VAULT__*."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer bapi_... (token do /credenciais)",
              "title": "Authorization"
            },
            "description": "Bearer bapi_... (token do /credenciais)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Listar Pix Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/pix/{txid}": {
      "get": {
        "tags": [
          "pix"
        ],
        "summary": "Consultar",
        "description": "`vencimento=true` consulta cobv (/cobv/{txid}); default cob.",
        "operationId": "consultar_pix__txid__get",
        "parameters": [
          {
            "name": "txid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Txid"
            }
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "vencimento",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Vencimento"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Provider",
              "default": "c6"
            }
          },
          {
            "name": "X-Bank-Credentials",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Credenciais do banco (JSON em base64) — só memória, nunca persistidas. Fallback: cofre VAULT__*.",
              "title": "X-Bank-Credentials"
            },
            "description": "Credenciais do banco (JSON em base64) — só memória, nunca persistidas. Fallback: cofre VAULT__*."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer bapi_... (token do /credenciais)",
              "title": "Authorization"
            },
            "description": "Bearer bapi_... (token do /credenciais)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PixCobrancaOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "pix"
        ],
        "summary": "Revisar",
        "description": "Revisa a cobrança (PATCH BACEN): valor, solicitacaoPagador, calendario...",
        "operationId": "revisar_pix__txid__patch",
        "parameters": [
          {
            "name": "txid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Txid"
            }
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "vencimento",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Vencimento"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Provider",
              "default": "c6"
            }
          },
          {
            "name": "X-Bank-Credentials",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Credenciais do banco (JSON em base64) — só memória, nunca persistidas. Fallback: cofre VAULT__*.",
              "title": "X-Bank-Credentials"
            },
            "description": "Credenciais do banco (JSON em base64) — só memória, nunca persistidas. Fallback: cofre VAULT__*."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer bapi_... (token do /credenciais)",
              "title": "Authorization"
            },
            "description": "Bearer bapi_... (token do /credenciais)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Campos"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PixCobrancaOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bolepix": {
      "post": {
        "tags": [
          "bolepix"
        ],
        "summary": "Criar",
        "description": "Emite Bolepix (boleto + QR Pix EVP). Reenvio com o mesmo\nexternal_reference_id devolve a cobrança existente (idempotente no banco).",
        "operationId": "criar_bolepix_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer bapi_... (token do /credenciais)",
              "title": "Authorization"
            },
            "description": "Bearer bapi_... (token do /credenciais)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BolepixIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CobrancaOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bolepix/{external_reference_id}": {
      "get": {
        "tags": [
          "bolepix"
        ],
        "summary": "Consultar",
        "description": "Consulta o Bolepix pelo external_reference_id (26 chars A-Z0-9).",
        "operationId": "consultar_bolepix__external_reference_id__get",
        "parameters": [
          {
            "name": "external_reference_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "External Reference Id"
            }
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Provider",
              "default": "c6"
            }
          },
          {
            "name": "X-Bank-Credentials",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Credenciais do banco (JSON base64) — só memória.",
              "title": "X-Bank-Credentials"
            },
            "description": "Credenciais do banco (JSON base64) — só memória."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer bapi_... (token do /credenciais)",
              "title": "Authorization"
            },
            "description": "Bearer bapi_... (token do /credenciais)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CobrancaOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "bolepix"
        ],
        "summary": "Cancelar",
        "description": "Cancela o Bolepix (409 enquanto a CIP processa o registro).",
        "operationId": "cancelar_bolepix__external_reference_id__delete",
        "parameters": [
          {
            "name": "external_reference_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "External Reference Id"
            }
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Provider",
              "default": "c6"
            }
          },
          {
            "name": "X-Bank-Credentials",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Credenciais do banco (JSON base64) — só memória.",
              "title": "X-Bank-Credentials"
            },
            "description": "Credenciais do banco (JSON base64) — só memória."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer bapi_... (token do /credenciais)",
              "title": "Authorization"
            },
            "description": "Bearer bapi_... (token do /credenciais)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CobrancaOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bolepix/{external_reference_id}/pdf": {
      "get": {
        "tags": [
          "bolepix"
        ],
        "summary": "Pdf",
        "description": "PDF do Bolepix em base64.",
        "operationId": "pdf_bolepix__external_reference_id__pdf_get",
        "parameters": [
          {
            "name": "external_reference_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "External Reference Id"
            }
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Provider",
              "default": "c6"
            }
          },
          {
            "name": "X-Bank-Credentials",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Credenciais do banco (JSON base64) — só memória.",
              "title": "X-Bank-Credentials"
            },
            "description": "Credenciais do banco (JSON base64) — só memória."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer bapi_... (token do /credenciais)",
              "title": "Authorization"
            },
            "description": "Bearer bapi_... (token do /credenciais)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CobrancaOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BolepixIn": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "provider": {
            "$ref": "#/components/schemas/Provider",
            "default": "c6"
          },
          "account_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Account Config"
          },
          "bolepix": {
            "$ref": "#/components/schemas/BolepixCobranca"
          },
          "credentials": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credentials"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "bolepix"
        ],
        "title": "BolepixIn"
      },
...
      "CobrancaOut": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id",
            "description": "Id da cobrança no banco (nosso número/txid)"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "linha_digitavel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linha Digitavel"
          },
          "codigo_barras": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codigo Barras"
          },
          "pix_copia_cola": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pix Copia Cola",
            "description": "PIX copia-e-cola (EMV), quando híbrido"
          },
          "pdf_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Base64",
            "description": "PDF do boleto em base64, quando disponível"
          },
          "raw": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw",
            "description": "Resposta crua do banco (debug)"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "CobrancaOut",
        "description": "Resposta normalizada — mesmo shape para qualquer provider."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PixCobrancaIn": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id",
            "description": "Identificador do tenant",
            "examples": [
              "empresa_123"
            ]
          },
          "provider": {
            "$ref": "#/components/schemas/Provider",
            "description": "Provider REST com Pix dinâmico (brcobrança não suporta).",
            "default": "c6"
          },
          "account_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Account Config",
            "description": "Blob por provider. c6: {chave_pix, ...}"
          },
          "pix": {
            "$ref": "#/components/schemas/PixCobranca"
          },
          "credentials": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credentials",
            "description": "Credenciais do banco enviadas NO REQUEST (client_id, client_secret, pfx_base64, pfx_password). Só em memória; fallback: cofre VAULT__*."
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "pix"
        ],
        "title": "PixCobrancaIn"
      },
      "PixCobrancaOut": {
        "properties": {
          "txid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Txid"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "pix_copia_cola": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pix Copia Cola",
            "description": "Payload EMV (copia-e-cola)"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location",
            "description": "URL do payload (loc.location)"
          },
          "expira_em": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expira Em"
          },
          "raw": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "PixCobrancaOut",
        "description": "Resposta normalizada da cobrança Pix."
      },
      "Provider": {
        "type": "string",
        "enum": [
          "brcobranca",
          "c6",
          "sicoob"
        ],
        "title": "Provider",
        "description": "Provider de cobrança. `brcobranca` = offline/CNAB (engine Ruby)."
      },
...
      "Status": {
        "type": "string",
        "enum": [
          "registrado",
          "pendente",
          "liquidado",
          "baixado",
          "expirado",
          "erro"
        ],
        "title": "Status",
        "description": "Status normalizado da cobrança (igual para qualquer banco)."
      },
...
    }
  },
...
}